Update loader scripts to work for GPDB
Prior to this commit, the generator script for loading census data
into PostGIS failed while executing the function. This commit updates
the loader script generation to work with GPDB.
The significant change that was made as part of this commit is to remove
inheritance: since the lookup tables are set to REPLICATED, and in GPDB
you cannot inherit a replicated table, instead of inheriting the base
table and creating state tables from it, we now create tables like the
base table and insert data into both the base as well as the underlying
table. We need to insert data into both these tables as inherit ensures
that it reads data from the child tables too, but now since they are 2
separate tables, we need to insert the relevant data into it, for the
geocoder functions to read from.